+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   `Readline` context created
 - Made sure to release initially `libreadline` allocated buffers to
   prevent one-time memory leak
+- Implemented `Default` trait for `Readline` struct
 
 
 0.1.0